fix: 修复错误的寄存器域定义；删除未使用的属性

+++ ldc1614/src/lib.rs
@@ -43,13 +43,13 @@ impl<const ADDR: u8> Ldc<ADDR> {
             Channel::One => self.register.data1_msb.read(i2c, DATA_MSB::data),
             Channel::Two => self.register.data2_msb.read(i2c, DATA_MSB::data),
             Channel::Three => self.register.data3_msb.read(i2c, DATA_MSB::data),
-        } << 16
-            | match ch {
+        } as u32) << 16
+            | (match ch {
                 Channel::Zero => self.register.data0_lsb.read(i2c, DATA_LSB::data),
                 Channel::One => self.register.data1_lsb.read(i2c, DATA_LSB::data),
                 Channel::Two => self.register.data2_lsb.read(i2c, DATA_LSB::data),
                 Channel::Three => self.register.data3_lsb.read(i2c, DATA_LSB::data),
-            }) as u32)
+            } as u32))
     }
     pub fn defaule_config<I2C: i2c::I2c>(&self, i2c: &mut I2C, ch: Channel) -> Result<()> {
         use bitmap::{
